all but the last exercise complete. these exercises were harder and are taking longer. but I feel i understand how authorization works, and how to manipulate it in using express tools

+++ part4/controllers/blogs.js
@@ -3,14 +3,6 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
-const getTokenFrom = request => {
-    const authorization = request.get('authorization')  
-    if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-        return authorization.substring(7)  
-    }  
-    return null
-}
-
 blogsRouter.get('/', (request, response) => {
     Blog
       .find({})
@@ -21,30 +13,27 @@ blogsRouter.get('/', (request, response) => {
   
 blogsRouter.post('/', async (request, response, next) => {
     const body = request.body
-    const token = getTokenFrom(request)
-    const decodedToken = jwt.verify(token, process.env.SECRET)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
-    if (!token || !decodedToken.id) {
+    if (!request.token || !decodedToken.id) {
         return response.status(401).json({ error: 'token missing or invalid' })
     } else {
+        const user = await User.findById(decodedToken.id)
+        console.log(user)
 
+        const blog = new Blog({
+          title: body.title,
+          author: body.author,
+          url: body.url,
+          likes: body.likes,
+          user: user
+        })
 
-      const user = await User.findById(decodedToken.id)
-      console.log(user)
-
-      const blog = new Blog({
-        title: body.title,
-        author: body.author,
-        url: body.url,
-        likes: body.likes,
-        user: user
-      })
-
-      const savedBlog = await blog.save()
-      user.blogs = user.blogs.concat(savedBlog._id)
-      await user.save()
-      
-      response.status(201).json(savedBlog)
+        const savedBlog = await blog.save()
+        user.blogs = user.blogs.concat(savedBlog._id)
+        await user.save()
+        
+        response.status(201).json(savedBlog)
     }
 })
 
@@ -66,8 +55,24 @@ blogsRouter.put('/:id', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+    const body = request.body
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!request.token || !decodedToken.id) {
+
+        return response.status(401).json({ error: 'token missing or invalid' })
+    } 
+
+    const blog = await Blog.findById(request.params.id)
+    
+    if (blog.user.toString() === decodedToken.id) {
+
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    } else {
+
+        response.status(401).json({ error: 'only the creator of this data can remove it '})
+    }
 })
 
 module.exports = blogsRouter
